complete rewrite with a heavy focus on simplicity and organization
Code is now organized strictly between
  1. Server code: src/*.ts
  2. HTML templates: components/*.ts
  3. Client code: static/*

The server code is rewritten with an emphasis on improving message
passing between the client/server. HTMX is no longer being used in
favor of a simpler JSON-only approach over a WebSocket. Chat messages
are still treated and stored as raw HTML, but will be sent through a
JSON object over the WebSocket and placed into the DOM using pure
JavaScript, rather than happening automatically with HTMX. This makes
message passing overall less convoluted, as there isn't HTMX data and
JSON data going through the same channel.

The HTML templates are greatly simplified, with less overall components
and less options given to the components. This is done so that the new
components focus strictly on the layout of the DOM, and not much else.
Modification to the DOM will rather happen in pure JavaScript on the
client, triggered by events received from the WebSocket.

The client's JavaScript code has been completely moved out of the HTML
template and into its own file. This is where the bulk of the processing
now happens. Simple events with small updates from the WebSocket are
processed, and DOM modifications are made accordingly; unlike previously
where SSR was used to pre-render the sidebar and send the *entire*
user list to the client every time somebody's status changed. This new
approach will also allow client-side filtering and sorting of the user
list, something previously infeasible with SSR.

Lastly, a constants.ts file was also added to simplify configuration
through environment variables.